0x0F-function_pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -6,9 +6,10 @@
  * @size: size of array
  * @action: function pointer
  */
+
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned long i;
 
 	for (i = 0; i < size; i++)
 		action(array[i]);
